dev panel tab for container messages
