+++ apps/web/package.json
@@ -29,6 +29,8 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.46.2",
+    "react-json-tree": "^0.18.0",
+    "react-json-view": "^1.21.3",
     "react-syntax-highlighter": "^15.5.0",
     "uuid": "^9.0.0",
     "zustand": "^4.4.3"
